refactor(analysis): extract collect stage to stages/collect.py
Verbatim move of collect_raw_data() from deep_analysis_pipeline.py to
src/finwiz/analysis/stages/collect.py. The pipeline re-exports the
function via `from finwiz.analysis.stages.collect import collect_raw_data`
so all callers remain unaffected. AnalysisContext is imported under
TYPE_CHECKING to avoid a circular import (same pattern as _helpers.py).
